fix: move EventBusRemoteWatcher to services tier (architecture boundary)
EventBusRemoteWatcher wraps EventBusProtocol which is services-tier.
Move it from core/remote_watcher.py to services/event_bus/remote_watcher.py
to satisfy the five-tier architecture boundary check.
